moved some functions from coverage.js to text.js
