+++ scripts/helpers/text.js
@@ -1,0 +1,57 @@
+/**
+ * @typedef {Object} Line
+ * @property {string} value
+ * @property {number} index
+ * @property {number} start
+ * @property {number} end
+ */
+
+/**
+ * @param {Array<Line>} lines
+ * @param {number} offset
+ * @return {Line}
+ */
+function getLineAtOffset(lines, offset) {
+  let lower = 0;
+  let upper = lines.length - 2;
+
+  while (lower < upper) {
+    const middle = (lower + (upper - lower) / 2) | 0;
+
+    if (offset < lines[middle].start) {
+      upper = middle - 1;
+    } else if (offset >= lines[middle + 1].start) {
+      lower = middle + 1;
+    } else {
+      lower = middle;
+      break;
+    }
+  }
+
+  return lines[lower];
+}
+
+exports.getLineAtOffset = getLineAtOffset;
+
+/**
+ * @param {string} input
+ * @return {Array<Line>}
+ */
+function parseLines(input) {
+  let offset = 0;
+
+  return input.split("\n").map((value, index) => {
+    const line = {
+      value,
+      index,
+      start: offset,
+      end: offset + value.length
+    };
+
+    offset = line.end + 1;
+
+    return line;
+  });
+}
+
+exports.parseLines = parseLines;
